Address race condition in process termination routines
- Added wait_until_completion(process_id, timeout) helper function

+++ rosbag2_test_common/include/rosbag2_test_common/process_execution_helpers_unix.hpp
@@ -94,33 +94,63 @@ ProcessHandle start_execution(const std::string & command)
   return process_id;
 }
 
+/// @brief Force to stop process with signal if it's currently running
+/// @param process_id Process ID
+/// @param signum Signal to use for stopping process. The default is SIGINT.
+/// @param timeout Timeout in fraction of seconds
 void stop_execution(
   const ProcessHandle & process_id,
   int signum = SIGINT,
-  std::chrono::seconds timeout = std::chrono::seconds(10))
+  std::chrono::duration<double> timeout = std::chrono::seconds(10))
 {
-  EXPECT_NE(kill(process_id, signum), -1) << "Failed to send signal " << signum <<
-    " to process: " << process_id;
-  int status = EXIT_FAILURE;
+  if (kill(process_id, 0) == 0) {
+    // If process is still running, then send signal to stop it and check return code
+    EXPECT_NE(kill(process_id, signum), -1) << "Failed to send signal " << signum <<
+      " to process: " << process_id;
+    int status = EXIT_FAILURE;
+    pid_t wait_ret_code = 0;
+    std::chrono::steady_clock::time_point const start = std::chrono::steady_clock::now();
+    // Wait for process to finish with timeout
+    while (wait_ret_code == 0 && std::chrono::steady_clock::now() - start < timeout) {
+      std::this_thread::sleep_for(std::chrono::milliseconds(30));
+      // WNOHANG - wait for processes without causing the caller to be blocked
+      wait_ret_code = waitpid(process_id, &status, WNOHANG);
+    }
+    if (wait_ret_code == 0) {
+      std::cerr << "Testing process " << process_id << " hangout. Killing it with SIGKILL \n";
+      kill(process_id, SIGKILL);
+    }
+    // Make sure that the process does execute without issues before it is killed by
+    // the user in the test or, in case it runs until completion, that it has correctly executed.
+    EXPECT_NE(wait_ret_code, -1);
+    EXPECT_EQ(wait_ret_code, process_id);
+    EXPECT_EQ(WIFEXITED(status), true) << "status = " << status;
+    EXPECT_EQ(WIFSIGNALED(status), false) << "Process terminated by signal: " << WTERMSIG(status);
+    EXPECT_EQ(WEXITSTATUS(status), EXIT_SUCCESS) << "status = " << status;
+  }
+}
+
+/// @brief Wait for process to finish with timeout
+/// @param process_id Process ID
+/// @param timeout Timeout in fraction of seconds
+/// @return true if process has finished during timeout and false if timeout was reached and
+/// process is still running
+bool wait_until_completion(
+  const ProcessHandle & process_id,
+  std::chrono::duration<double> timeout = std::chrono::seconds(5))
+{
   pid_t wait_ret_code = 0;
+  int status = EXIT_FAILURE;
   std::chrono::steady_clock::time_point const start = std::chrono::steady_clock::now();
   // Wait for process to finish with timeout
   while (wait_ret_code == 0 && std::chrono::steady_clock::now() - start < timeout) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(30));
+    std::this_thread::sleep_for(std::chrono::milliseconds(10));
     // WNOHANG - wait for processes without causing the caller to be blocked
     wait_ret_code = waitpid(process_id, &status, WNOHANG);
   }
-  if (wait_ret_code == 0) {
-    std::cerr << "Testing process " << process_id << " hangout. Killing it with SIGKILL \n";
-    kill(process_id, SIGKILL);
-  }
-  // Make sure that the process does execute without issues before it is killed by
-  // the user in the test or, in case it runs until completion, that it has correctly executed.
   EXPECT_NE(wait_ret_code, -1);
-  EXPECT_EQ(wait_ret_code, process_id);
-  EXPECT_EQ(WIFEXITED(status), true) << "status = " << status;
-  EXPECT_EQ(WIFSIGNALED(status), false) << "Process terminated by signal: " << WTERMSIG(status);
-  EXPECT_EQ(WEXITSTATUS(status), EXIT_SUCCESS) << "status = " << status;
+  EXPECT_EQ(wait_ret_code, process_id) << "status = " << status;
+  return wait_ret_code != 0;
 }
 
 #endif  // ROSBAG2_TEST_COMMON__PROCESS_EXECUTION_HELPERS_UNIX_HPP_
